+++ RELEASE_NOTES.md
@@ -12,14 +12,14 @@ Breaking changes can still occur in minor versions and patches, though, if the a
 
 ## 5.2
 
-This version adds support for `russian` and `ukranian`.
+This version adds support for `russian` and `ukrainian`.
 
 This version also bumps the package's Swift version to 5.5 to enable extended DocC support.
 
 ### 🌐 New locales
 
 * 🇷🇺 Russian
-* 🇺🇦 Ukranian
+* 🇺🇦 Ukrainian
 
 ### 🐛 Bug fixes